refactor: use hero instead of colums for a title and a form
