+++ campaign/blocks/hero/hero.css
@@ -1,40 +1,158 @@
-/* block specific CSS goes here */
-
-main .hero-container > div {
-    max-width: unset;
-}
-  
-main .hero-container {
-    padding: 0;
-}
-  
-main .hero {
-    position: relative;
-    padding: 2rem;
-    min-height: 300px;
+main .section>div.hero-wrapper {
+    max-width: 1400px;
 }
 
 main .hero h1 {
-    max-width: 768px;
-    margin-left: auto;
-    margin-right: auto;
-    color: white;
+    margin: 0;
 }
 
-
-main .hero picture {
-    position: absolute;
-    z-index: -1;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    object-fit: cover;
-    box-sizing: border-box;
+main .hero h2 {
+    margin-top: 40px;
 }
 
-main .hero img {
-    object-fit: cover;
+main .hero h3 {
+    margin-bottom: 0;
+}
+
+main .hero>div {
+    display: flex;
+    flex-direction: column;
+    padding: 24px 0;
+    align-items: center;
+    justify-content: space-between;
+}
+
+main .hero>div:nth-child(even) {
+    display: flex;
+    flex-direction: column-reverse;
+}
+
+main .hero .form-container {
+    box-shadow: 0px 1px 6px rgb(34 34 34 / 10%);
+    flex-basis: 100%;
     width: 100%;
-    height: 100%;
+}
+
+main .hero .form-container h3 {
+    font-size: 32px;
+    font-weight: normal;
+}
+
+main .hero .form-container .error {
+    color: #eb0510;
+}
+
+main .hero p {
+    font-size: 19px;
+}
+
+main .hero .form-container input,
+main .hero .form-container select {
+    background-color: white;
+    border-radius: 4px;
+    border: 1px solid rgb(193, 193, 195);
+    color: rgb(29, 31, 37);
+    font-size: 14px;
+    line-height: 21px;
+    padding: 16px;
+
+}
+
+main .hero .form-container input:hover,
+main .hero .form-container select:hover {
+    border-color: rgb(90, 92, 96);
+}
+
+main .hero .form-container .error input,
+main .hero .form-container .error select {
+    margin-bottom: 0;
+}
+
+main .hero .form-container label {
+    font-family: var(--font-family-heading);
+    font-size: 11px;
+    font-weight: 700;
+    letter-spacing: 0.2px;
+    text-transform: uppercase;
+}
+
+main .hero .form-container span.error {
+    font-size: 14px;
+    margin-top: 8p;
+}
+
+main .hero .form-container .form-checkbox-wrapper {
+    display: flex;
+    margin: 16px 0;
+}
+
+main .hero .form-container .form-checkbox-wrapper input[type='checkbox'] {
+    display: none;
+}
+
+main .hero .form-container .form-checkbox-wrapper input[type='checkbox']+label:before {
+    float: left;
+    margin-top: 5px;
+    margin-right: 5px;
+    height: 30px;
+    width: 30px;
+    content: ' ';
+    background: url('./checkbox.svg');
+    background-size: contain;
+}
+
+main .hero .form-container .form-checkbox-wrapper input[type='checkbox']:checked+label:before {
+    background: url('./checkbox-checked.svg');
+    background-size: contain;
+    content: ' ';
+}
+
+main .hero .form-container .form-checkbox-wrapper label {
+    font-size: 14px;
+    color: #FAFAFA;
+    font-weight: normal;
+    text-transform: none;
+    font-family: var(--font-family-body);
+    display: block;
+
+}
+
+@media (min-width: 768px) {
+
+    main .hero>div,
+    main .hero>div:nth-child(even) {
+        flex-direction: unset;
+        text-align: left;
+    }
+
+    main .hero>div>div {
+        width: 100%;
+    }
+
+    main .hero>div>div:first-child {
+        margin-right: 16px;
+    }
+
+    main .hero .form-container {
+        flex-basis: 390px;
+        flex-shrink: 0;
+        flex-grow: 0;
+    }
+}
+
+@media (min-width: 1024px) {
+    main .hero>div>div:last-child {
+        margin-left: 16px;
+    }
+
+    main .hero .form-container .form-short {
+        width: 48%;
+        display: inline-flex;
+        align-items: normal;
+        flex-direction: column;
+    }
+
+    main .hero .form-container .form-short:nth-child(1n+2) {
+        margin-left: 4%;
+    }
 }
